Add commented out code from StackOverflow
Their example is to use ioutil.ReadDir() to build a list of files
and then use sort.Slice() and what appears to be a Time object
to sort that list by modification time.

This could work for our first prototype, but we'll also want to offer
the ability to sort by pattern/filename (e.g, based on the datestamps
used in some auto-generated files).