Add NOTICE file
Use the existing license file and the two license
files for `jessevdk/go-flags` and `sirupsen/logrus`,
to craft a first draft of the NOTICE.txt file.

---

Request:

If anyone spots anything out of place, in the wrong
format, etc., please submit a Pull Request or
open an issue noting the problems with this file.

Thanks in advance.

refs #14