Split main package into multiple subpackages
This is inspired by #76, but also because I've known for
a while that this might eventually be needed. I'm hoping
to better isolate code from the main package in the
hope/belief that this will make setting up future tests
easier and more reliable.

As part of this work, the majority of the existing utility
functions had to be renamed to export them for use elsewhere
and a few functions had to be updated to accept a Config
object in order to receive application-wide settings. This
Config object also (now) bundles a `*logrus.Logger` for
shared use across the application.

I don't know Golang well enough at this point to know
whether this is a terrible idea. If you're reading this
and *you* know the answer, please reach out and share.

refs #87, #76