Update GitHub Action Workflow name
Update overall workflow name to make it distinct from the
job(s) that run as part of the workflow. The intent is to
make it easier to select a "group" of status checks via
the repo branch protection rule(s) and to better communicate
what the overall goal of the workflow is.

refs #127