Clarify config source precedence behavior
Attempt to make clear how multiple configuration
sources are handled. AFAIK, it is standard practice
to have later configuration sources (or those higher
in precedence) completely override settings from
earlier in the load order, regardless of whether the
existing value is a default value.

This implementation (currently) only overrides earlier
settings if the incoming configuration has non-default
values specified. This behavior is subject to change if
others find it too confusing or inflexible for normal use.

refs #156