Prototyping panic behavior after post-merge config validation failure
Need to refine; should NewConfig() handle panicing or should this
behavior be handled by main() function?

In short, would I expect the go-arg, go-toml and other config
packages to immediately panic when problems are encountered
or would I expect them to return the error "to me" in main() so
I could consider further for final behavior?

I think the latter ...