Remove AppMetadata field guards in Config.String()
In short, the guard checks do not appear to be needed.
Whether this is due to recent work to remove the support
for overriding the AppMetadata fields via external config
sources I do not recall.

By removing this code we rely entirely on the Getter
methods defined for the purpose handling AppMetadata field
values.

This commit also updates the `TestLoadConfigFileTemplate`
test to use those same Getter methods instead of relying on
the old behavior.

refs #195