Restore missing version output since v0.7.0
The version string shown in both the Help and Version flag
output has been missing since v0.7.0 when support for
setting configuration information via config files was
added and config merging precedence was reworked.

We provide the missing information by directly initializing
the appropriate metadata field for the flags config struct
prior to calling `arg.MustParse()`. This ensures that
the version info is available when the `Version()` method
is called by the `alexflint/go-arg` package.

refs GH-240