Improve HasMatching* doc comments
Note that lack of user-specified values causes the checks to
immediately flag the specified filename as eligible for
removal. This is a potential "gotcha" as one might assume
that a lack of a clear match would indicate that the file
should be skipped.

While this logic likely needs further evaluation, this minor
doc update should help make the intent clearer for the
short-term.