Update build tags for Go 1.17 compatibility
golangci-lint v1.42.1 introduced changes to the `gofmt` linter.
These changes appear to match the same ones introduced with
Go 1.17 which expect build tags to be in a specific format
OR include an additional leading set of build tags with
the `//go:` prefix.

This commit adds this additional build tag entry to both
os-specific variants of the `logging_*.go` file.

refs GH-350